refactor: standardize API client methods
